Remove v1 code 🎉
