+++ src/commands/explorer/groupBy.ts
@@ -5,7 +5,6 @@
 
 import { IActionContext } from "@microsoft/vscode-azext-utils";
 import { QuickPickItem } from "vscode";
-import { getArmTagKeys } from "../../utils/azureUtils";
 import { localize } from "../../utils/localize";
 import { settingUtils } from "../../utils/settingUtils";
 
@@ -24,8 +23,8 @@ async function groupBy(context: IActionContext, setting: string): Promise<void> 
     await settingUtils.updateGlobalSetting('groupBy', setting);
 }
 
-async function getQuickPicks(context: IActionContext): Promise<QuickPickItem[]> {
-    return Array.from((await getArmTagKeys(context))).map(key => { return { label: key } })
+async function getQuickPicks(_context: IActionContext): Promise<QuickPickItem[]> {
+    throw new Error('Getting arm tag keys is not implemented yet');
 }
 
 export enum GroupBySettings {
