--- src/resolvers/ShallowResourceResolver.ts
@@ -1,32 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-*  Copyright (c) Microsoft Corporation. All rights reserved.
-*  Licensed under the MIT License. See License.txt in the project root for license information.
-*--------------------------------------------------------------------------------------------*/
-
-import { ISubscriptionContext } from "@microsoft/vscode-azext-utils";
-import { AppResource, AppResourceResolver, ResolvedAppResourceBase } from "@microsoft/vscode-azext-utils/hostapi";
-import { TreeItemCollapsibleState } from "vscode";
-import { getAzureExtensions } from "../AzExtWrapper";
-import { BuiltinResolver } from "./BuiltinResolver";
-
-/**
- * This resolver acts as a "placeholder" resolver when no extension is known for the resource type.
- * It changes nothing about the node. When the extension activates, it will supply the real resolver.
- */
-class ShallowResourceResolver implements AppResourceResolver, BuiltinResolver {
-    public readonly resolverKind = 'builtin';
-
-    public resolveResource(_subContext: ISubscriptionContext, _resource: AppResource): ResolvedAppResourceBase {
-        return {
-            commandId: 'azureResourceGroups.viewProperties',
-            initialCollapsibleState: TreeItemCollapsibleState.None
-        };
-    }
-
-    public matchesResource(resource: AppResource): boolean {
-        return !getAzureExtensions().some(azExt => azExt.matchesResourceType(resource));
-    }
-}
-
-export const shallowResourceResolver = new ShallowResourceResolver();
-
